+++ src/autoharness/__init__.py
@@ -3,6 +3,9 @@
 from autoharness.models import (
     AgentTrace,
     AnalysisResult,
+    AutoFixAttemptFeedback,
+    AutoFixAttemptResult,
+    AutoFixPhase,
     AutoFixPipelineResult,
     EvaluationResult,
     EvolutionPipelineResult,
@@ -18,6 +21,9 @@ __all__ = [
     "AgentTrace",
     "AnalysisResult",
     "AutoHarness",
+    "AutoFixAttemptFeedback",
+    "AutoFixAttemptResult",
+    "AutoFixPhase",
     "AutoFixPipelineResult",
     "EvaluationResult",
     "EvolutionPipelineResult",
@@ -27,4 +33,4 @@ __all__ = [
     "RepairPipelineResult",
     "SkillRecommendationResult",
 ]
-__version__ = "0.5.0"
+__version__ = "0.6.0"